+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## 1.2.0
+* `ImageAssetConfigurationBase#slice` を追加
+
+## 1.1.0
+* @akashic/pdi-types@1.3.0 に追従
+* `VectorImageAsset` の設定を追加
+
 ## 1.0.1
 * `LoadGameConfigurationFunc` の引数順を修正